+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for `menu-bar` mode
 - Add support for background refresh parameter `interval`
+- Add `Open Command File` command
 
 ## 0.6.1